+++ src/Router.tsx
@@ -1,25 +1,26 @@
-import React, { useEffect } from "react";
-import { Routes, Route, useLocation } from "react-router-dom";
+import React from "react";
+import { createBrowserRouter } from "react-router-dom";
 import MainPage from "pages/MainPage";
 import DetailPage from "pages/DetailPage";
-import Header from "components/layout/Header";
+import Root from "Root";
 
-const Router = () => {
-  const location = useLocation();
+const router = createBrowserRouter([
+  {
+    path: "/",
+    id: "root",
+    // errorElement: <Error />,
+    element: <Root />,
+    children: [
+      {
+        path: "/",
+        element: <MainPage />,
+      },
+      {
+        path: "/videos",
+        element: <DetailPage />,
+      },
+    ],
+  },
+]);
 
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [location]);
-
-  return (
-    <>
-      <Header />
-      <Routes>
-        <Route path="/" element={<MainPage />} />
-        <Route path="/videos" element={<DetailPage />} />
-      </Routes>
-    </>
-  );
-};
-
-export default Router;
+export default router;
